requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
billiard==3.6.4.0
black==23.3.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2.23
django-embed-video==1.4.8
django-environ==0.10.0
django-etc==1.4.0
django-hitcount==1.3.5
django-map-widgets==0.4.1
django-paypal==2.0
django-phonenumber-field==7.1.0
django-pwa==1.1.0
django-redis==5.3.0
django-taggit==4.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
djangorestframework-gis==1.0
fcm-django==2.0.0
firebase-admin==6.1.0
frozenlist==1.3.3
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
kombu==5.2.4
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
phonenumberslite==8.13.14
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.10.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
tomli==2.0.1
tqdm==4.65.0
twilio==8.3.0
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
